+++ pkg/ingester/lifecycle_test.go
@@ -2,8 +2,11 @@ package ingester
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"math"
+	"net/http"
+	"net/http/httptest"
 	"testing"
 	"time"
 
@@ -64,12 +67,12 @@ func TestIngesterRestart(t *testing.T) {
 	config := defaultIngesterTestConfig()
 	clientConfig := defaultClientTestConfig()
 	limits := defaultLimitsTestConfig()
-	config.LifecyclerConfig.SkipUnregister = true
+	config.LifecyclerConfig.UnregisterFromRing = false
 
 	{
 		_, ingester := newTestStore(t, config, clientConfig, limits, nil)
 		time.Sleep(100 * time.Millisecond)
-		// doesn't actually unregister due to skipUnregister: true
+		// Doesn't actually unregister due to UnregisterFromRing: false.
 		require.NoError(t, services.StopAndAwaitTerminated(context.Background(), ingester))
 	}
 
@@ -80,7 +83,7 @@ func TestIngesterRestart(t *testing.T) {
 	{
 		_, ingester := newTestStore(t, config, clientConfig, limits, nil)
 		time.Sleep(100 * time.Millisecond)
-		// doesn't actually unregister due to skipUnregister: true
+		// Doesn't actually unregister due to UnregisterFromRing: false.
 		require.NoError(t, services.StopAndAwaitTerminated(context.Background(), ingester))
 	}
 
@@ -89,6 +92,32 @@ func TestIngesterRestart(t *testing.T) {
 	test.Poll(t, 100*time.Millisecond, 1, func() interface{} {
 		return testutils.NumTokens(config.LifecyclerConfig.RingConfig.KVStore.Mock, "localhost", ring.IngesterRingKey)
 	})
+}
+
+func TestIngester_ShutdownHandler(t *testing.T) {
+	for _, unregister := range []bool{false, true} {
+		t.Run(fmt.Sprintf("unregister=%t", unregister), func(t *testing.T) {
+			config := defaultIngesterTestConfig()
+			clientConfig := defaultClientTestConfig()
+			limits := defaultLimitsTestConfig()
+			config.LifecyclerConfig.UnregisterFromRing = unregister
+			_, ingester := newTestStore(t, config, clientConfig, limits, nil)
+
+			// Make sure the ingester has been added to the ring.
+			test.Poll(t, 100*time.Millisecond, 1, func() interface{} {
+				return testutils.NumTokens(config.LifecyclerConfig.RingConfig.KVStore.Mock, "localhost", ring.IngesterRingKey)
+			})
+
+			recorder := httptest.NewRecorder()
+			ingester.ShutdownHandler(recorder, nil)
+			require.Equal(t, http.StatusNoContent, recorder.Result().StatusCode)
+
+			// Make sure the ingester has been removed from the ring even when UnregisterFromRing is false.
+			test.Poll(t, 100*time.Millisecond, 0, func() interface{} {
+				return testutils.NumTokens(config.LifecyclerConfig.RingConfig.KVStore.Mock, "localhost", ring.IngesterRingKey)
+			})
+		})
+	}
 }
 
 func TestIngesterChunksTransfer(t *testing.T) {